requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
python-dotenv==1.0.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
yarl==1.8.2
